requirements.txt
streamlit
pandas
thefuzz
python-Levenshtein
plotly
gspread
google-auth-oauthlib